requirements.txt
python-dotenv
faster-whisper
torch
numpy
requests
reportlab
openai
python-telegram-bot==20.7
fastapi
uvicorn
python-multipart